NMS-12603: backport circleci from foundation-2018

+++ tools/packages/minion/minion.spec
@@ -121,37 +121,12 @@ rm -rf %{buildroot}
 
 %install
 
-export OPTS_MAVEN="-Daether.connector.basic.threads=1 -Daether.connector.resumeDownloads=false"
-export OPTS_SKIP_TESTS="-DskipITs=true -Dmaven.test.skip.exec=true"
-export OPTS_SKIP_TARBALL="-Dbuild.skip.tarball=true"
-export OPTS_ASSEMBLIES="-Passemblies"
-export OPTS_PROFILES="-Prun-expensive-tasks"
-export COMPILE_PROJECTS="org.opennms.features.minion:repository,org.opennms.features.minion:core-impl,org.opennms.features.minion:core-repository,org.opennms.features.minion:shell,org.opennms.features.minion:core-features,org.opennms.features.minion:core-parent,org.opennms.features.minion:core-api,org.opennms.features.minion.shell:org.opennms.features.minion.shell.provision,org.opennms.features.minion.shell:org.opennms.features.minion.shell.poller,org.opennms.features.minion:org.opennms.features.minion.shell,org.opennms.features.minion:minion-parent,org.opennms.features.minion.container:karaf,org.opennms.features.minion.container:extender,org.opennms.features.minion.container:features,org.opennms.features.minion:container-parent,org.opennms.features.minion.container:scv,org.opennms.features.minion.heartbeat:org.opennms.features.minion.heartbeat.producer,org.opennms.features.minion:org.opennms.features.minion.heartbeat,org.opennms.features.minion.heartbeat:org.opennms.features.minion.heartbeat.common,org.opennms.features.minion.heartbeat:org.opennms.features.minion.heartbeat.consumer"
-export ASSEMBLY_PROJECTS=":org.opennms.assemblies.minion"
-
+export EXTRA_ARGS=""
 if [ "%{enable_snapshots}" = 1 ]; then
-	OPTS_ENABLE_SNAPSHOTS="-Denable.snapshots=true"
-	OPTS_UPDATE_POLICY="-DupdatePolicy=always"
+	EXTRA_ARGS="-s"
 fi
 
-# always build the root POM, just to be sure inherited properties/plugin/dependencies are right
-./compile.pl -N $OPTS_SKIP_TESTS $OPTS_SKIP_TARBALL $OPTS_SETTINGS_XML $OPTS_ENABLE_SNAPSHOTS $OPTS_UPDATE_POLICY -Dinstall.version="%{version}-%{release}" -Ddist.name="%{buildroot}" -Dopennms.home="%{instprefix}" install
-if [ "%{skip_compile}" = 1 ]; then
-	echo "=== SKIPPING FULL COMPILE ==="
-	echo "Projects: ${ASSEMBLY_PROJECTS}"
-	./compile.pl $OPTS_MAVEN $OPTS_SKIP_TESTS $OPTS_SKIP_TARBALL $OPTS_SETTINGS_XML $OPTS_ENABLE_SNAPSHOTS $OPTS_UPDATE_POLICY $OPTS_PROFILES $OPTS_ASSEMBLIES -Dinstall.version="%{version}-%{release}" -Ddist.name="%{buildroot}" \
-		-Dopennms.home="%{instprefix}" \
-		--projects "${ASSEMBLY_PROJECTS}" \
-		install
-else
-	# get the full list of minion projects to build
-	echo "=== RUNNING COMPILE ==="
-	echo "Projects: ${COMPILE_PROJECTS},${ASSEMBLY_PROJECTS}"
-	./compile.pl $OPTS_MAVEN $OPTS_SKIP_TESTS $OPTS_SKIP_TARBALL $OPTS_SETTINGS_XML $OPTS_ENABLE_SNAPSHOTS $OPTS_UPDATE_POLICY $OPTS_PROFILES $OPTS_ASSEMBLIES -Dinstall.version="%{version}-%{release}" -Ddist.name="%{buildroot}" \
-		-Dopennms.home="%{instprefix}" \
-		--projects "${COMPILE_PROJECTS},${ASSEMBLY_PROJECTS}" --also-make \
-		install
-fi
+tools/packages/minion/create-minion-assembly.sh $EXTRA_ARGS
 
 # Extract the minion assembly
 mkdir -p %{buildroot}%{minioninstprefix}
@@ -220,7 +195,14 @@ ROOT_INST="${RPM_INSTALL_PREFIX0}"
 # Clean out the data directory
 if [ -d "${ROOT_INST}/data" ]; then
     find "$ROOT_INST/data/"* -maxdepth 0 -name tmp -prune -o -print0 | xargs -0 rm -rf
-    find "$ROOT_INST/data/tmp/"* -maxdepth 0 -name README -prune -o -print0 | xargs -0 rm -rf
+    if [ -d "${ROOT_INST}/data/tmp"  ]; then
+        find "$ROOT_INST/data/tmp/"* -maxdepth 0 -name README -prune -o -print0 | xargs -0 rm -rf
+    fi
+fi
+
+# Clean out .m2 directory
+if [ -d "${ROOT_INST}/.m2" ]; then
+   rm -rf "${ROOT_INST}/.m2"
 fi
 
 # Generate an SSH key if necessary
